refactor(shape-plugin): split progress panel state and cache actions helpers
